finished refactoring etc the functions.
We really only need to use apply_pbcs and give it a center point.

For a useful one we give it a center point which is the centroid of a
grouping of the protein and ligand or pair.

The function we are interested though at a higher level here is
`recenter_pair` which will recenter a protein ligand complex