refactor and remove apply_pbc from group centering
body="""
There were issues with the use of applying the PBCs to freshly
recentered complexs. The motivation was to set them at a center point,
although the cutting off of things due to apply_PBCs here was not
necessary and results in information loss.

If you want to apply the pbcs after centering it is very easy to do
so.

For distances that require the use of PBCs this might be useful.

However, for things like RMSD, SASA, and visualization it is much more
useful to always have whole structures, since they don't even regard
the PBCs anyhow.

"""

changes = [:changes/fixation, :changes/accretion, :changes/rename,
:changes/replaced]

domain = [:domain/src]