updating docstrings
[body]

Removed a bunch of those annoying things in the functions where you
could control the number of return arguments. Just pick which ones you
want...

This really is a renaming... but. Also you don't have to specify all
those bools in the arguments so that is a relaxation

changes = [:changes/removal, :changes/relaxed]