removes superimpose_traj method
[body]

THis was not idiomatic with the rest of the code and since it just
does a list comprehension I thought it is a bad idea to bloat the
namespace with an x_traj method for every method.

It will be better to have a apply_traj method that applies any
operation to multiple trajs.

changes = [:changes/removal, :changes/clarity]
domain = [:domain/src, :domain/docs]